UUID: avoid shifting by more than bit width by casting to size_t
uu[i] gets promoted to an int (32 bit) but it shifted by up to 56 bits
on 64 bit platforms which results in undefined behaviour

this caused gcc 9.3 to emit code that miscalculated the hash in loops

cast uu[i] to size_t (64 bit on 64 bit platforms) to avoid undefined
behaviour

fixes #453

+++ src/util/uuid.hpp
@@ -38,7 +38,7 @@ public:
     {
         size_t r = 0;
         for (size_t i = 0; i < 16; i++) {
-            r ^= uu[i] << ((i % sizeof(size_t)) * 8);
+            r ^= ((size_t)uu[i]) << ((i % sizeof(size_t)) * 8);
         }
         return r;
     }
